Extract button actions from main file for readability Added center on map expanded button
